Implement some benchmark running, fix llvm function value, fix tests

+++ dev/floyd/floyd_runtime/floyd_corelib.cpp
@@ -328,7 +328,7 @@ extern const std::string k_corelib_builtin_types_and_constants = R"(
 
 	struct seq_t {
 		string str
-		size_t pos
+		int pos
 	}
 
 	struct text_t {
